fix: discard changes in package.json before committing
This fixes a possible crash when a package.json has devDependencies that are not sorted alphabetically. Note that this only applies when devDependencies are not locked, for instance when using npm@3

`npm install --save-dev` sorts dependencies in the package.json, so when a lockfile is not changed by this script, it tries to commit nothing, which fails.