feat: handle monorepos
BREAKING CHANGE: This now works with monorepos, it’s a big refactor,
that’s why we make it a new semver-major release when it lands, just
so that existing users won’t get tripped up by little changes.

However, the design goal here is to be as backwards compatible as
possible.