fix: handle amend commits when trying to find a lockfile commit.
Amend commits might obscure the commit message, which we were
using previously to determine whether greenkeeer-lockfile has
run before.

This commit changes the detection to looking at the files changed
for the list of commits that are on the current branch. It stops
the run, if it includes at least one of these lockfiles:

- package-lock.json
- npm-shrinkwrap.json
- yarn.lock

Part of #174